requirements.txt
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
numpy==1.23.1
yapf==0.40.1
timm
transformers
mxnet-mkl==1.6.0
torchtoolbox
addict
pandas
albumentations==0.4.6